+++ config/database.js
@@ -1,13 +1,6 @@
-import mongoose from 'mongoose';
-
-const options = {
-    useUnifiedTopology: true,
-    useNewUrlParse: true
-}
+import mongoose from 'mongoose'
 
 mongoose.set('strictQuery', false)
-
-
-mongoose.connect(process.env.MONGO_URI,options)
+mongoose.connect(process.env.MONGO_URI)
     .then(() => console.log('database connected'))
     .catch(err => console.log(err))
