--- models/data/chapter.data.js
@@ -1,9 +1,0 @@
-import 'dotenv/config.js'
-import '../../config/database.js'
-import {Chapter} from '../Chapter.models.js'
-
- let chapters = [
-
- ]
-
- Chapter.insertMany(chapters)
